@@ -15,45 +15,48 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 # In iOS, build-name is used as CFBundleShortVersionString while build-number used as CFBundleVersion.
 # Read more about iOS versioning at
 # https://developer.apple.com/library/archive/documentation/General/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html
-version: 1.0.0+1
+version: 1.1.0+1
 
 environment:
-  sdk: ">=2.17.0 <3.0.0"
+  # sdk: ">=2.17.0 <3.0.0"
+  sdk: ">=2.19.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
+  # excel: ^4.0.2
   uuid: ^4.0.0
   cupertino_icons: ^1.0.0
   english_words: ^4.0.0
   json_annotation: ^4.0.1
-  intl: ^0.17.0
+  intl: ^0.19.0
   equatable: ^2.0.3
-  flutter_bloc: ^7.2.0
+  flutter_bloc: ^8.1.3
   firebase_core: ^2.5.0
   firebase_auth: ^4.2.6
   cloud_firestore: ^4.3.2
   firebase_storage: ^11.0.11
-  universal_html: ^2.0.8
+  # universal_html: ^2.0.8
+  flutter_widget_from_html: ^0.14.11
   firebase_storage_web: ^3.3.20
   firebase_messaging: ^14.2.4
-  xml: ^5.3.0
-  flutter_login: ^4.1.1
-  adaptive_dialog: ^1.8.0+1
-  flutter_html: ^2.2.1
+  xml: ^6.5.0
+  flutter_login: ^5.0.0
+  adaptive_dialog: ^2.0.0
+  # flutter_html: ^2.2.1
   flutter_rating_bar: ^4.0.1
   uni_links: ^0.5.1
-  responsive_framework: ^0.2.0
+  responsive_framework: ^1.1.1
   # data_table_2: ^2.2.1
   # syncfusion_flutter_datepicker: ^19.2.44
-  json_serializable: ^4.1.3
-  pluto_grid: ^2.10.3
+  json_serializable: ^6.7.1
+  # pluto_grid: ^7.0.2
   data_tables: ^1.0.3
-  grouped_list: ^4.0.0
-  flutter_native_splash: ^1.2.0
+  grouped_list: ^5.1.2
+  flutter_native_splash: ^2.3.9
   fluttertoast: ^8.1.2
-  file_picker: ^5.2.5
+  file_picker: ^6.1.1
   enum_to_string: ^2.0.1
 
 dev_dependencies:
